[CI] Split xpu worker restart plugin into pytest_configs/ package
Extract the install logic into xpu_smi_install.py and the xpu-smi snapshot
poller into xpu_mem_monitor.py, leaving xpu_worker_restart.py as a thin
pytest plugin. Move all three under .github/scripts/pytest_configs/ and
update PYTHONPATH and the -p plugin name in the Linux and Windows UT
workflows accordingly.
