+++ .github/workflows/_linux_ut.yml
@@ -69,9 +69,9 @@ jobs:
               -e ZE_AFFINITY_MASK
       env:
         AGENT_TOOLSDIRECTORY: /tmp/xpu-tool
-        PYTHONPATH: ${{ github.workspace }}/.github
-        PYTEST_ADDOPTS: -v --timeout 600 --timeout_method=thread --max-worker-restart 1000000
-                        -p scripts.xpu_worker_restart
+        PYTHONPATH: ${{ github.workspace }}/.github/scripts/pytest_configs
+        PYTEST_ADDOPTS: -v --timeout 600 --timeout_method=thread
+                        --max-worker-restart 1000000 -p xpu_worker_restart
                         --only-rerun 'crashed while running' --reruns 2
                         --dist worksteal ${{ needs.runner.outputs.pytest_extra_args }}
     steps:
@@ -109,8 +109,10 @@ jobs:
     runs-on: ${{ needs.runner.outputs.runner_id }}
     env:
       AGENT_TOOLSDIRECTORY: /tmp/xpu-tool
-      PYTHONPATH: ${{ github.workspace }}/.github
-      PYTEST_ADDOPTS: -v --timeout 3600 --timeout_method=thread -n 1 --max-worker-restart 10000 -p scripts.xpu_worker_restart
+      PYTHONPATH: ${{ github.workspace }}/.github/scripts/pytest_configs
+      PYTEST_ADDOPTS: -v -n 1
+                      --timeout 3600 --timeout_method=thread
+                      --max-worker-restart 10000 -p xpu_worker_restart
     steps:
       - name: Checkout torch-xpu-ops
         uses: actions/checkout@v4
